test(notifications): split NotificationServiceTests by feature
Replace Unit/NotificationServiceTests.cs with per-feature files under
Features/Notifications/<Op>/, sharing a NotificationServiceTestFixture base.
Cross-cutting tests (NotifierTests, TestBackgroundJobs) remain in Unit/.
